requirements.txt
# Requirements for Agentic Research System
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
streamlit>=1.40.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
